Move cfg_mgt_app.go -> transformer/cfg_mgmt_xfmr.go Remove host_comm.go
